// src/gql/graphql.ts
export const ProductImageFragmentDoc = new TypedDocumentString(`
    fragment ProductImage on Asset {
  url
  fileName
}
    `, {"fragmentName":"ProductImage"}) as unknown as TypedDocumentString<ProductImageFragment, unknown>;
export const SingleProductItemFragmentDoc = new TypedDocumentString(`
    fragment SingleProductItem on Product {
  id
  name
  price
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
    fragment ProductImage on Asset {
  url
  fileName
}`, {"fragmentName":"SingleProductItem"}) as unknown as TypedDocumentString<SingleProductItemFragment, unknown>;
export const CartFragmentDoc = new TypedDocumentString(`
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    product {
      ...SingleProductItem
    }
  }
}
    fragment SingleProductItem on Product {
  id
  name
  price
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
fragment ProductImage on Asset {
  url
  fileName
}`, {"fragmentName":"Cart"}) as unknown as TypedDocumentString<CartFragment, unknown>;
export const SingleProductSizeColorVariantFragmentDoc = new TypedDocumentString(`
    fragment SingleProductSizeColorVariant on ProductSizeColorVariant {
  id
  name
}
    `, {"fragmentName":"SingleProductSizeColorVariant"}) as unknown as TypedDocumentString<SingleProductSizeColorVariantFragment, unknown>;
// ...
export const ProductListItemFragmentDoc = new TypedDocumentString(`
    fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
    fragment ProductImage on Asset {
  url
  fileName
}`, {"fragmentName":"ProductListItem"}) as unknown as TypedDocumentString<ProductListItemFragment, unknown>;
export const ReviewItemFragmentDoc = new TypedDocumentString(`
    fragment ReviewItem on Review {
  id
  name
  headline
  email
  content
  rating
}
    `, {"fragmentName":"ReviewItem"}) as unknown as TypedDocumentString<ReviewItemFragment, unknown>;
export const CartAddProductDocument = new TypedDocumentString(`
    mutation CartAddProduct($cartId: ID!, $productId: ID!, $total: Int!) {
  createOrderItem(
    data: {quantity: 1, total: $total, order: {connect: {id: $cartId}}, product: {connect: {id: $productId}}}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartAddProductMutation, CartAddProductMutationVariables>;
export const CartCreateOrderDocument = new TypedDocumentString(`
    mutation CartCreateOrder {
  createOrder(data: {total: 0}) {
    ...Cart
  }
}
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    product {
      ...SingleProductItem
    }
  }
}
fragment SingleProductItem on Product {
  id
  name
  price
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
fragment ProductImage on Asset {
  url
  fileName
}`) as unknown as TypedDocumentString<CartCreateOrderMutation, CartCreateOrderMutationVariables>;
export const CartDeleteProductDocument = new TypedDocumentString(`
    mutation CartDeleteProduct($itemId: ID!) {
  deleteOrderItem(where: {id: $itemId}) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartDeleteProductMutation, CartDeleteProductMutationVariables>;
export const CartGetByIdDocument = new TypedDocumentString(`
    query CartGetById($id: ID!) {
  order(where: {id: $id}, stage: DRAFT) {
    ...Cart
  }
}
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    product {
      ...SingleProductItem
    }
  }
}
fragment SingleProductItem on Product {
  id
  name
  price
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
fragment ProductImage on Asset {
  url
  fileName
}`) as unknown as TypedDocumentString<CartGetByIdQuery, CartGetByIdQueryVariables>;
export const CartSetProductQuantityDocument = new TypedDocumentString(`
    mutation CartSetProductQuantity($itemId: ID!, $quantity: Int!, $total: Int!) {
  updateOrderItem(
    where: {id: $itemId}
    data: {quantity: $quantity, total: $total}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartSetProductQuantityMutation, CartSetProductQuantityMutationVariables>;
export const CartUpsertProductDocument = new TypedDocumentString(`
    mutation CartUpsertProduct($orderItemId: ID, $cartId: ID!, $productId: ID!, $total: Int!, $quantity: Int!) {
  upsertOrderItem(
    upsert: {create: {quantity: 1, total: $total, order: {connect: {id: $cartId}}, product: {connect: {id: $productId}}}, update: {quantity: $quantity, total: $total}}
    where: {id: $orderItemId}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartUpsertProductMutation, CartUpsertProductMutationVariables>;
export const CategoriesGetByCategorySlugDocument = new TypedDocumentString(`
    query CategoriesGetByCategorySlug($slug: String!) {
  categories(where: {slug: $slug}) {
    id
    name
    slug
  }
}
    `) as unknown as TypedDocumentString<CategoriesGetByCategorySlugQuery, CategoriesGetByCategorySlugQueryVariables>;
export const CategoriesGetListDocument = new TypedDocumentString(`
    query CategoriesGetList {
  categories {
    id
    name
    slug
  }
}
    `) as unknown as TypedDocumentString<CategoriesGetListQuery, CategoriesGetListQueryVariables>;
export const CollectionsGetCollectionBySlugDocument = new TypedDocumentString(`
    query CollectionsGetCollectionBySlug($slug: String!) {
  collections(where: {slug: $slug}) {
    id
    name
    slug
  }
}
    `) as unknown as TypedDocumentString<CollectionsGetCollectionBySlugQuery, CollectionsGetCollectionBySlugQueryVariables>;
export const CollectionsGetListDocument = new TypedDocumentString(`
    query CollectionsGetList {
  collections {
    id
    name
    slug
    image {
      url
      fileName
    }
  }
}
    `) as unknown as TypedDocumentString<CollectionsGetListQuery, CollectionsGetListQueryVariables>;
export const OrderUpdateCheckoutStripeIdDocument = new TypedDocumentString(`
    mutation OrderUpdateCheckoutStripeId($id: ID, $stripeCheckoutId: String) {
  updateOrder(data: {stripeCheckoutId: $stripeCheckoutId}, where: {id: $id}) {
    id
  }
}
    `) as unknown as TypedDocumentString<OrderUpdateCheckoutStripeIdMutation, OrderUpdateCheckoutStripeIdMutationVariables>;
export const ProductGetByIdDocument = new TypedDocumentString(`
    query ProductGetById($id: ID!) {
  product(where: {id: $id}) {
    ...SingleProductItem
  }
}
    fragment SingleProductItem on Product {
  id
  name
  price
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}
fragment ProductImage on Asset {
  url
  fileName
}`) as unknown as TypedDocumentString<ProductGetByIdQuery, ProductGetByIdQueryVariables>;
export const ProductGetReviewsRatingDocument = new TypedDocumentString(`
    query ProductGetReviewsRating($id: ID!) {
  reviewsConnection(where: {product: {id: $id}}) {
    edges {
      node {
        rating
      }
    }
    aggregate {
      count
    }
  }
}
    `) as unknown as TypedDocumentString<ProductGetReviewsRatingQuery, ProductGetReviewsRatingQueryVariables>;
export const ProductGetVariantsListDocument = new TypedDocumentString(`
    query ProductGetVariantsList($id: ID!) {
  product(where: {id: $id}) {
    variants {
      ... on ProductColorVariant {
        ...SingleProductColorVariant
      }
      ... on ProductSizeColorVariant {
        ...SingleProductSizeColorVariant
      }
      ... on ProductSizeVariant {
        ...SingleProductSizeVariant
      }
    }
  }
}
    fragment SingleProductSizeColorVariant on ProductSizeColorVariant {
  id
  name
}
fragment SingleProductColorVariant on ProductColorVariant {
  id
  name
}
fragment SingleProductSizeVariant on ProductSizeVariant {
  id
  name
}`) as unknown as TypedDocumentString<ProductGetVariantsListQuery, ProductGetVariantsListQueryVariables>;
export const ProductUpdateAverageRatingDocument = new TypedDocumentString(`
    mutation ProductUpdateAverageRating($averageRating: Int!, $id: ID!) {
  updateProduct(data: {averageRating: $averageRating}, where: {id: $id}) {
    id
    averageRating
  }
  publishProduct(to: PUBLISHED, where: {id: $id}) {
    id
  }
}
    `) as unknown as TypedDocumentString<ProductUpdateAverageRatingMutation, ProductUpdateAverageRatingMutationVariables>;
export const ProductsCountDocument = new TypedDocumentString(`
    query ProductsCount {
  productsConnection {
    aggregate {
      count
    }
  }
}
    `) as unknown as TypedDocumentString<ProductsCountQuery, ProductsCountQueryVariables>;
export const ProductsCountByCategorySlugDocument = new TypedDocumentString(`
    query ProductsCountByCategorySlug($slug: String!) {
  categories(where: {slug: $slug}) {
    id
    name
    slug
    products {
      id
    }
  }
}
    `) as unknown as TypedDocumentString<ProductsCountByCategorySlugQuery, ProductsCountByCategorySlugQueryVariables>;
export const ProductGetListDocument = new TypedDocumentString(`
    query ProductGetList($limit: Int!, $offset: Int!, $orderBy: ProductOrderByInput) {
  productsConnection(first: $limit, skip: $offset, orderBy: $orderBy) {
    products: edges {
      node {
        ...ProductListItem
      }
    }
    pageInfo {
      hasNextPage
      hasPreviousPage
    }
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductGetListQuery, ProductGetListQueryVariables>;
export const ProductsGetListByCategorySlagDocument = new TypedDocumentString(`
    query ProductsGetListByCategorySlag($slag: String!, $limit: Int!, $offset: Int!, $orderBy: ProductOrderByInput) {
  productsConnection(
    where: {categories_some: {slug: $slag}}
    first: $limit
    skip: $offset
    orderBy: $orderBy
  ) {
    products: edges {
      node {
        ...ProductListItem
      }
    }
    pageInfo {
      hasNextPage
      hasPreviousPage
    }
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListByCategorySlagQuery, ProductsGetListByCategorySlagQueryVariables>;
export const ProductsGetListByCollectionSlagDocument = new TypedDocumentString(`
    query ProductsGetListByCollectionSlag($slag: String!) {
  products(where: {collections_some: {slug: $slag}}) {
    ...ProductListItem
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListByCollectionSlagQuery, ProductsGetListByCollectionSlagQueryVariables>;
export const ProductsGetListSearchDocument = new TypedDocumentString(`
    query ProductsGetListSearch($search: String!) {
  products(where: {_search: $search}) {
    ...ProductListItem
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListSearchQuery, ProductsGetListSearchQueryVariables>;
export const ReviewCreateDocument = new TypedDocumentString(`
    mutation ReviewCreate($id: ID!, $headline: String!, $name: String!, $email: String!, $content: String!, $rating: Int!) {
  createReview(
    data: {headline: $headline, name: $name, email: $email, content: $content, rating: $rating, product: {connect: {id: $id}}}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<ReviewCreateMutation, ReviewCreateMutationVariables>;